various changes
Some changes in index.html updating the text content of the page and form reformulation in style.css because this not work correctly. Additionally, added the correct images in the first part and third part of the main content. Finally, was changed the title of the page to "AutoLab" instead of only "landingpage".  That's it for today.